Add EdgeFlags enum for evaluation graph edge flags
Replaces the raw int Flags field on EvaluationGraph.Edge with a
[Flags] enum. Only bit 2 (0x04) has a known meaning: Invertible,
marking a bidirectional lookup connection. The remaining bits are
neutral placeholders (Bit0, Bit1, Bit3–Bit5) following the same
convention as NodeFlags.

+++ src/ACadSharp/IO/DWG/DwgStreamWriters/DwgObjectWriter.Objects.cs
@@ -1074,7 +1074,7 @@ internal partial class DwgObjectWriter : DwgSectionIO
 		for (int i = 0; i < edgesArr.Length; i++)
 		{
 			this._writer.WriteBitLong(i);
-			this._writer.WriteBitLong(edgesArr[i].Flags);
+			this._writer.WriteBitLong((int)edgesArr[i].Flags);
 			this._writer.WriteBitLong(edgesArr[i].TrackedCount);
 
 			this._writer.WriteBitLong(edgesArr[i].FromNodeIndex);
